+++ src/views/dashboard/related/components/LogTable/Index.vue
